+++ src/test/java/com/arkea/devops/sonar/xml/checks/DescriptionPatternCheckTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and 
  * limitations under the License.
  */
-package com.arkea.devops.sonar.xml;
+package com.arkea.devops.sonar.xml.checks;
 
 import static org.junit.Assert.assertEquals;
